% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRE_fns.R
\name{approximateLTRE}
\alias{approximateLTRE}
\title{Approximate LTRE analysis}
\usage{
approximateLTRE(Aobj, method = "random")
}
\arguments{
\item{Aobj}{An object containing all the population projection matrices to be
included in the analysis. It should either be a list, or a matrix where each
row is the column-wise vectorization of a matrix.

For fixed design, exactly 2 matrices must be provided, ordered as
\verb{[reference matrix, test matrix}]. For random design, any set of 2 or more
matrices can be provided. The set of matrices passed in must all have the
same dimensions.}

\item{method}{Either "random" or "fixed." The default behavior is "random."
See details for more information.}
}
\value{
A matrix of contributions to variance (random design) or difference
(fixed design) in lambda. Lambda is the asymptotic population growth rate,
defined as the largest eigenvalue of the population projection matrix.
}
\description{
Life Table Response Experiments (LTREs) are a method of comparative demographic
analysis. The purpose is to quantify how the difference or variance in vital
rates (stage-specific survival, growth, and fertility) amongst populations
contributes to difference or variance in the population growth rate, "lambda."
The equations and descriptions for the approximate methods of LTRE analysis
can be found in Caswell's 2001 textbook.
}
\details{
Lambda is the asymptotic population growth rate, defined as the
largest eigenvalue of the population projection matrix. A fixed design LTRE
decomposes the difference in lambda due to differences at each position of
the matrices. For a fixed design LTRE, exactly 2 matrices must be provided,
ordered as \verb{[reference matrix, test matrix}]. The matrix of contributions
returned from a approximate method fixed design LTRE will have the same shape
as the provided matrices.

A random design LTRE decomposes the variance in lambda due to variance and
covariance in the entries at each position in the matrices. For a random
design LTRE, at least 2 matrices must be provided. The matrix of
contributions returned from a approximate method random design LTRE will
include both first-order terms (due to variance) and interaction terms (due
to covariance). Therefore, if the provided matrix is 3x3, the matrix of
contributions will be 9x9 (the size of the variance-covariance matrix is the
square of the size of the original matrix). The contributions of variances
are found on the diagonal of the contribution matrix, and the contributions
of covariances are symmetric. So the contribution of covariance between two
vital rate parameters is the sum of the two corresponding off-diagonal terms.

The equations and descriptions for the approximate methods of LTRE analysis
can be found in Caswell's 2001 textbook.
}
\examples{
A1<- matrix(data=c(0,0.8,0, 0,0,0.7, 5,0,0.2), nrow=3, ncol=3)
A2<- matrix(data=c(0,0.9,0, 0,0,0.5, 4,0,0.3), nrow=3, ncol=3)
A3<- matrix(data=c(0,0.4,0, 0,0,0.6, 6,0,0.25), nrow=3, ncol=3)
cont_diff<- approximateLTRE(list(A1,A2), method='fixed') # contributions to the difference in lambda
# contributions to the variance of lambda
cont_var<- approximateLTRE(list(A1,A2,A3), method='random')
}
