% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRE_fns.R
\name{approximateLTRE_fixed}
\alias{approximateLTRE_fixed}
\title{Approximate LTRE analysis: fixed design}
\usage{
approximateLTRE_fixed(Aref, Atest)
}
\arguments{
\item{Aref}{The population projection matrix of the reference population.
Depending on the experimental or observational dataset, this may be the
control treatment, the first time period, or the unharvested population, etc.}

\item{Atest}{The population projection matrix of the test population.}
}
\value{
A matrix of contributions to the difference in lambda. Lambda is the
asymptotic population growth rate, defined as the largest eigenvalue of the
population projection matrix.
}
\description{
Approximate LTRE analysis: fixed design
}
\details{
Lambda is the asymptotic population growth rate, defined as the
largest eigenvalue of the population projection matrix. A fixed design LTRE
decomposes the difference in lambda due to differences at each position of
the matrices. For a fixed design LTRE, exactly 2 matrices must be provided,
ordered as \verb{[reference matrix, test matrix}]. The matrix of contributions
returned from an approximate method fixed design LTRE will have the same shape
as the provided matrices.

In some cases, it may not be obvious how to identify the reference and the
test matrix. The sum of contributions will be approximately equal to the
observed difference in lambda between these two matrices, evaluated as
lambda(Atest) - lambda(Aref). In cases where it doesn't 'matter' which way
you, as a user, input these matrices, it is important to understand how to
interpret positive and negative contributions.

The equations and descriptions for the approximate methods of LTRE analysis
can be found in Caswell's 2001 textbook.
}
\examples{
A1<- matrix(data=c(0,0.8,0, 0,0,0.7, 5,0,0.2), nrow=3, ncol=3)
A2<- matrix(data=c(0,0.9,0, 0,0,0.5, 4,0,0.3), nrow=3, ncol=3)
cont_diff<- approximateLTRE(list(A1,A2), method='fixed') # contributions to the difference in lambda
}
