% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fns.R
\name{cov_matrix}
\alias{cov_matrix}
\title{The variance-covariance matrix for a set of population projection matrices}
\usage{
cov_matrix(Aobj)
}
\arguments{
\item{Aobj}{A list of matrix population models, which must all have the same
dimensions.}
}
\value{
The variance-covariance matrix for the vital rates of the population
projection matrices. If the dimensions of each matrix in \code{Aobj} are n-by-n,
then the variance-covariance matrix will have dimensions n^2-by-n^2.
}
\description{
Calculate the variance-covariance matrix for the vital rates of a series of
population projection matrices.
}
\examples{
A1<- matrix(data=c(0,0.8,0, 0,0,0.7, 5,0,0.2), nrow=3, ncol=3)
A2<- matrix(data=c(0,0.9,0, 0,0,0.5, 4,0,0.3), nrow=3, ncol=3)
A3<- matrix(data=c(0,0.4,0, 0,0,0.6, 6,0,0.25), nrow=3, ncol=3)
covmat<- cov_matrix(list(A1,A2,A3))
}
